requirements.txt
absl-py==2.1.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
APScheduler==3.10.4
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
DAWG-Python==0.7.2
decorator==5.1.1
distro==1.8.0
docopt==0.6.2
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.103.2
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2023.12.2
gast==0.5.4
google-auth-oauthlib==0.4.6
google-auth==2.29.0
google-pasta==0.2.0
gradio-client==0.8.1
grpcio==1.62.2
h11==0.14.0
h5py==3.11.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.20.3
idna==3.4
ipython==8.24.0
jedi==0.19.1
joblib==1.4.0
Keras-Preprocessing==1.1.2
keras==2.8.0
kiwisolver==1.4.5
libclang==18.1.1
markdown-it-py==3.0.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
multidict==6.0.4
namex==0.0.8
navec==0.10.0
numpy==1.26.2
oauthlib==3.2.2
openai==1.1.1
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
pip==22.3.1
prompt-toolkit==3.0.43
protobuf==3.20.0
pure-eval==0.2.2
py-cord==2.4.1
pyasn1-modules==0.4.0
pyasn1==0.6.0
pycparser==2.21
pydantic-core==2.10.1
pydantic==2.4.2
pygments==2.17.2
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy2==0.9.1
PyNaCl==1.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-telegram-bot==20.6
pytz==2023.3.post1
PyYAML==6.0.1
requests-oauthlib==2.0.0
requests==2.31.0
rich==13.7.1
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
setuptools==65.5.1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.8.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow==2.8.0
termcolor==2.4.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.4.0
toxicity==0.30
tqdm==4.66.1
traitlets==5.14.3
typing-extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.6
wcwidth==0.2.13
websockets==11.0.3
werkzeug==3.0.2
wheel==0.38.4
wrapt==1.16.0
yarl==1.9.4
